feature(leaderboard): can see all scores on leaderboard
