+++ client/components/index.js
@@ -7,5 +7,5 @@ export {default as GameSelect} from './GameSelect'
 export {default as GameInit} from './GameInit'
 export {default as GameItem} from './GameItem'
 export {default as Home} from './Home'
-export {default as Leaderboard} from './Leaderboards'
+export {default as Leaderboards} from './Leaderboards'
 export {default as NoMatch} from './NoMatch'
