+++ client/routes.js
@@ -8,7 +8,7 @@ import {
   Game2,
   Game3,
   GameSelect,
-  Leaderboard,
+  Leaderboards,
   NoMatch
 } from './components'
 
@@ -21,7 +21,7 @@ class Routes extends Component {
       <Switch>
         <Route exact path="/" component={Home} />
         <Route path="/select" component={GameSelect} />
-        <Route path="/leaderboard" component={Leaderboard} />
+        <Route path="/leaderboard" component={Leaderboards} />
         <Route
           exact
           path="/game1"
