+++ server/routes/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router()
 module.exports = router
 
-router.use('/clockleaderboard', require('./clockleaderboard'))
-router.use('/normalleaderboard', require('./normalleaderboard'))
-router.use('/bombleaderboard', require('./bombleaderboard'))
+router.use('/clockGame', require('./clockleaderboard'))
+router.use('/normalGame', require('./normalleaderboard'))
+router.use('/bombGame', require('./bombleaderboard'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
